+++ .github/workflows/ci-pipeline.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-22.04
     steps:
       - uses: actions/checkout@v3
-      - run: pip install ruff
+      - run: pip install ruff==0.1.6
       - run: scripts/check-format.sh
 
   check-static-analyzer:
@@ -20,7 +20,7 @@ jobs:
     runs-on: ubuntu-22.04
     steps:
       - uses: actions/checkout@v3
-      - run: pip install ruff
+      - run: pip install ruff==0.1.6
       - run: cmake --preset unix
       - run: scripts/check-lint.sh
