+++ images/plots/cornell-box-0-2-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1fedac54c2c403d1623cf288479716fad1ee944e94bb3264627fd1e0775cedd0
-size 359150
+oid sha256:24c6bcdd81b7a055b42e2789a8352ed8db69bed420189254c49897b3f683afaa
+size 359334
