+++ images/plots/cornell-box-0-2-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b6fe47b2a1a58a2d102e16f185d5454c74061093fc8db780bc4d8108da97e3b1
-size 360161
+oid sha256:c925f08298e95ee4a9af9a483b4a5975e6874461d249e04734687bc8046cc577
+size 360375
