+++ images/plots/cornell-box-1-2-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:50e677ea9ab888fa2a2137ed18dc54deb636082b56f781f31e411c7f1590a940
-size 915779
+oid sha256:0cc13cd89d280a4d1fce8f25f4742445a9f83ae528bda15974dcd9607911fd95
+size 915770
