+++ images/plots/cornell-box-1-2-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:194ff645048d42ae6295c158fe97a3e4a574d1a0ac1476b33dcb685fb334e8a2
-size 916418
+oid sha256:5c04ec66268eb9db9d0eea236b4449e0b9211366a81bff0d8909a335fc4ee7eb
+size 916449
