+++ images/plots/cornell-box-1-32-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0b2ec21d84626be8494e9f02f445f07695f38f5ac14ade04b1724feb2dd5ea42
-size 871835
+oid sha256:d3784f77c1c68350f46bd20cc1f6c1e79472c2e6311a2f59e2ccdfba53370a91
+size 871754
