+++ images/plots/cornell-box-1-32-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e3acf86123e50a80fd30df951a15a89b00fb9f3c827aa1e83981bce0f2472222
-size 872708
+oid sha256:d7214cac2f7b7b2ba3f7d3b0d66b5d8aca07f590e7a08a91f8848486d4107791
+size 872644
