+++ images/plots/error-filter-space-plot-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:26d919cfb1d475f2e9ac39393874f686f4ec76377040f8750e1727ef30da0d41
-size 491930
+oid sha256:c6f52f2710a76e9dd6e89a5cb04cae0f7e7b94bd2b7d530696277c91d76618e1
+size 633312
