+++ images/plots/error-filter-space-plot-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:72c868dd8bd524eddf2638fa296266064eb677ef9c0fa48a4b8fedbd832fad6c
-size 518328
+oid sha256:578ec21593c5ad94049156505e04cfcbce2fd7eb7fcedc14a9087ce61846e919
+size 639070
