+++ images/plots/error-filter-time-plot-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4c2c679a6e99a6f29f20afd56ca92f51f403736578406a923ba73c88034a7b2a
-size 217188
+oid sha256:26f1fb3ba10f20251213ab1f41c976957ff041454262710ac249c7dbe8c9a48c
+size 299964
