+++ images/plots/error-filter-time-plot-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:edbd35262be44e5521edeedbe68491c49b66cc4ee5950cec32b5b33cf91d2976
-size 223868
+oid sha256:809b2e672ac116fe539ce1af0827fd26ae737def074b02019c6186a58348196e
+size 303650
