+++ images/plots/error-plot-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:8dc71ef4a9c8b38fbd586e9952267576fcfddcb89e209ff77712f1b7bceb14e5
-size 308621
+oid sha256:b033ad32a02fb7466fe8c20c74da14c9c4944c3d21f309692a2e10782d213ee8
+size 353902
