+++ images/plots/error-plot-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6483af711612687a4a5d75336cbcecac78d87f732100026e3a695479578bbecb
-size 316507
+oid sha256:c1eedf15bc3eaa3cd33b5b61a95e398ddb4dfe4b5ca9ef9008a15d037c2659fb
+size 363507
