+++ images/plots/mc-qmc-compared-dark.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:69273cf37dfd10aeab994cb61a569c1d0ca5aab60b899ada021723e3fd80a350
+size 203834
