+++ images/plots/mc-qmc-compared-light.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:23273b3c650fc5f74e8ff8d1bdff8dd94434e1f7033a0bfc878dd18669e0f278
+size 212194
