+++ images/plots/pair-plot-lattice-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:ec6fce072f69eff34cf584f30e8def6bf0de04e2bc18c22a32a6612f832e9b86
-size 1313558
+oid sha256:7719380c4ae7f358f36d398d6c719b3381ae9db73dd82b7a34ef1591eea376c4
+size 1416271
