+++ images/plots/pair-plot-lattice-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:cc34372170d79d9dc9989aa35cea122b30583a1b55390767bb7c626eb647a5bb
-size 1321224
+oid sha256:c6f9589245be397b65ae085259e6eaac10f48817880e435e1dafe7a4563a4a3a
+size 1422326
