+++ images/plots/pair-plot-pmj-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f371fc8b751fd2a23445b738ee74cb78c4a2e587fafc0a51ddddd2693209a419
-size 1550281
+oid sha256:2f2e1b89a300b0065d7560bf9c25e310ed57ee04b649c538f0b7d4a3f5521843
+size 1661971
