+++ images/plots/pair-plot-pmj-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6a55556cda724ffce5c0af8a4c4b208871e82cc568d0c3ee73e0b3995eace866
-size 1559015
+oid sha256:c593f593c237eb526d4eab2b5ac224c1522c44eaf1e88f473382ba8fe9645143
+size 1673768
