+++ images/plots/pair-plot-sobol-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5af8674f934e21475e4a3bb471086a7e54054b6b8b578523961b3b03d6f4a579
-size 1507967
+oid sha256:ef2a0c440fbc966ec070f8cce4ef347da857ac6bf66c7306a609c87690fcd055
+size 1619597
