+++ images/plots/pair-plot-sobol-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:db024cb98e69dfcab5651bf8677bb7ab5263a857608126f26410d6889b89de3f
-size 1515330
+oid sha256:d6f8dc84dedeeebd7d385ff5d7dac3495bb7685a89b2f43a8974a0aee6dcd942
+size 1631473
