+++ images/plots/performance-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f665727393474a5d07f20d7c28a14c3347bece84fcf13678ee07ec5af7247e5f
-size 66604
+oid sha256:4557c2bd0fa5972dff6814e970e22692dd6630644c4fc7f62ef9f1e444526ba7
+size 66804
