+++ images/plots/performance-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:dd06db51098319b42d68040c656acde79396dd495fb4f48d9476f90009d14c89
-size 70039
+oid sha256:423ac3ecfbb6dfbc4f88e477e8df004dd3baead586faaa02472f6f1ef6ac81d6
+size 70224
