+++ images/plots/zoneplate-plot-dark.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c7e2c01df93ef4e453094320e7c502e67a775a893c6dd66dcba9de87ba673830
-size 1545112
+oid sha256:2f87c86089b22477cf7515c28c5242aeb092aeb50b56eccf3704596dae44e09e
+size 1547634
