+++ images/plots/zoneplate-plot-light.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:92feba87f6e288958bece4b24d9c758f3af7221cdf8f7922523e4aaef0f9f80e
-size 1548677
+oid sha256:e10bdf3f983cfab7b89be430eecc1e4795f71a2f1c53d239ccce44594eead53a
+size 1550862
